Kirills work
git-svn-id: http://b035a0a.online-server.cloud/svn/AcfPublic@1423 7b0a5285-9844-ea4a-88f9-6912f79b293c

+++ Acf/Include/imath/CVarVector.cpp
@@ -52,12 +52,12 @@ void CVarVector::SetElementsFrom(const CVarVector& vector, double expansionValue
 
 bool CVarVector::EnsureElementsCount(int count, double value)
 {
-	int oldSize = m_elements.size();
+	size_t oldSize = m_elements.size();
 	if (count > oldSize){
 		m_elements.resize(count);
 
 		if (value != 0){
-			for (int i = oldSize; i < count; ++i){
+			for (size_t i = oldSize; i < count; ++i){
 				m_elements[i] = value;
 			}
 		}
